+++ common-service/src/main/java/com/webacademy/common/repositories/CourseRepository.java
@@ -1,9 +1,7 @@
-package com.webacademy.course.repository;
+package com.webacademy.common.repositories;
 
 import com.webacademy.common.entities.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface CourseRepository extends JpaRepository<Course,Long> {
