--- course-lecture-service/src/main/java/com/webacademy/course_lecture/repository/CourseRepo.java
@@ -1,7 +1,0 @@
-package com.webacademy.course_lecture.repository;
-
-import com.webacademy.common.entities.Course;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CourseRepo extends JpaRepository<Course, Long> {
-}
