+++ course-service/src/main/java/com/webacademy/course/config/CourseConfig.java
@@ -2,21 +2,15 @@ package com.webacademy.course.config;
 
 import com.webacademy.common.entities.Course;
 import com.webacademy.common.entities.CourseInformation;
-import com.webacademy.common.entities.CourseLecture;
 import com.webacademy.common.entities.Teacher;
-import com.webacademy.course.repository.CourseRepository;
+import com.webacademy.common.repositories.CourseRepository;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import javax.transaction.Transactional;
-import java.lang.reflect.Array;
 import java.time.LocalDateTime;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
 
 @Configuration @Transactional
 public class CourseConfig {
