+++ student-service/src/main/java/com/webacademy/student/config/StudentConfig.java
@@ -1,7 +1,7 @@
 package com.webacademy.student.config;
 
 import com.webacademy.common.entities.Student;
-import com.webacademy.student.repository.StudentRepository;
+import com.webacademy.common.repositories.StudentRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
